only checking options on the RPC level, removed unwanted code, will continue later
