+++ core/types/blob_tx_wrapper.go
@@ -381,11 +381,3 @@ func (txw *BlobTxWrapper) MarshalBinary(w io.Writer) error {
 func (txw BlobTxWrapper) EncodeRLP(w io.Writer) error {
 	return txw.Tx.EncodeRLP(w)
 }
-
-func (txw BlobTxWrapper) PutOptions(options *types2.TransactionConditions) {
-	txw.Tx.TransactionConditions = options
-}
-
-func (txw BlobTxWrapper) GetOptions() *types2.TransactionConditions {
-	return txw.Tx.TransactionConditions
-}
