+++ eth/stagedsync/stage_mining_exec.go
@@ -437,16 +437,6 @@ LOOP:
 			continue
 		}
 
-		// not prioritising conditional transaction, yet.
-		if options := txn.GetOptions(); options != nil {
-			if err := ValidateTransactionConditions(options, header, ibs, false); err != nil {
-				log.Trace("Dropping conditional transaction", "from", from, "hash", txn.Hash(), "reason", err)
-				txs.Pop()
-
-				continue
-			}
-		}
-
 		// Check whether the txn is replay protected. If we're not in the EIP155 (Spurious Dragon) hf
 		// phase, start ignoring the sender until we do.
 		if txn.Protected() && !chainConfig.IsSpuriousDragon(header.Number.Uint64()) {
@@ -506,30 +496,4 @@ func NotifyPendingLogs(logPrefix string, notifier ChainEventNotifier, logs types
 		return
 	}
 	notifier.OnNewPendingLogs(logs)
-}
-
-func ValidateTransactionConditions(tc *types2.TransactionConditions, header *types.Header, ibs *state.IntraBlockState, lengthCheckFlag bool) error {
-	// check block number range
-	if _, err := types2.BigIntIsWithinRange(header.Number, tc.BlockNumberMin, tc.BlockNumberMax); err != nil {
-		return &types2.TransactionConditionsValidationError{Message: "out of block range. err: " + err.Error()}
-	}
-
-	// check timestamp range
-	if _, err := types2.Uint64IsWithinRange(&header.Time, tc.TimestampMin, tc.TimestampMax); err != nil {
-		return &types2.TransactionConditionsValidationError{Message: "out of time range. err: " + err.Error()}
-	}
-
-	// check knownAccounts
-	if err := ibs.ValidateKnownAccounts(tc.KnownAccountStorageConditions); err != nil {
-		return &types2.KnownAccountsLimitExceededError{Message: "limit exceeded. err: number of slots/accounts in KnownAccountStorageConditions exceeds the limit of 1000"}
-	}
-
-	if lengthCheckFlag {
-		// check knownAccounts length (number of slots/accounts) should be less than 1000
-		if tc.KnownAccountStorageConditions.CountStorageEntries() >= 1000 {
-			return &types2.KnownAccountsLimitExceededError{Message: "limit exceeded. err: number of slots/accounts in KnownAccountStorageConditions exceeds the limit of 1000"}
-		}
-	}
-
-	return nil
 }
